requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
WTForms==2.3.3
requests==2.26.0
XlsxWriter==3.0.2
gunicorn==20.1.0
flask-caching==1.10.1
brotli-asgi==1.1.0
gevent==21.12.0
PyGithub==1.55 